+++ rtp-core/src/main/java/io/github/dailystruggle/rtp/common/configuration/enums/PerformanceKeys.java
@@ -15,6 +15,7 @@ public enum PerformanceKeys {
   effectParsing,
   biomeRecall,
   biomeRecallForced,
+  biomeWeighted,
   visitorEnabled,
   loginCacheEnabled,
   loginCacheCap,